+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 # RxDB Changelog
 
 <!-- CHANGELOG NEWEST -->
-
+- ADD [fastify adapter to RxServer](https://rxdb.info/rx-server.html)
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->